Handling GeometryTileWorker edge cases:
 - If redoLayout completes without a placementConfig, move to NeedPlacement
 - If setPlacementConfig called while in NeedPlacement, attemptPlacement
 - If setPlacementConfig called while in NeedLayout and there are not other dependencies, attemptPlacement

The idea with these changes is that waiting on the first placementConfig is essentially similar to a symbol dependency, and we need to make sure we do layout/placement in response to it being satisfied. However, we only want to trigger activity the first time, to avoid short circuiting the coalescing logic.

 - Don't re-trigger layout when symbol dependencies arrive unless all of our prior dependencies have been satisfied first

This avoids a case where we can have multiple "different" sets of dependencies in flight at the same time, and we could end up rendering with the wrong set.